+++ packages/media-parser/src/boxes/webm/get-ready-tracks.ts
@@ -1,7 +1,7 @@
 import type {Track} from '../../get-tracks';
-import {getTracksSegment} from '../../traversal';
 import {getTrack} from './get-track';
 import type {MainSegment} from './segments/all-segments';
+import {getTracksSegment} from './traversal';
 
 export const getTracksFromMatroska = (
 	segment: MainSegment,
