+++ packages/media-parser/src/test/av1.test.ts
@@ -1,10 +1,10 @@
 import {RenderInternals} from '@remotion/renderer';
 import {expect, test} from 'bun:test';
+import {getMoovBox, getTraks} from '../boxes/iso-base-media/traversal';
 import {trakBoxContainsVideo} from '../get-fps';
 import {getAv1CBox} from '../get-sample-aspect-ratio';
 import {parseMedia} from '../parse-media';
 import {nodeReader} from '../readers/from-node';
-import {getMoovBox, getTraks} from '../traversal';
 
 test('AV1 in MP4', async () => {
 	const parsed = await parseMedia({
